feat: footer bytes with capacity
