+++ src/puffin/src/file_format/writer/footer.rs
@@ -19,7 +19,7 @@ use snafu::ResultExt;
 
 use crate::blob_metadata::BlobMetadata;
 use crate::error::{Result, SerializeJsonSnafu};
-use crate::file_format::MAGIC;
+use crate::file_format::{MAGIC, MIN_FOOTER_SIZE};
 use crate::file_metadata::FileMetadataBuilder;
 
 /// Writer for the footer of a Puffin file.
@@ -43,10 +43,15 @@ impl FooterWriter {
 
     /// Serializes the footer to bytes
     pub fn into_footer_bytes(mut self) -> Result<Vec<u8>> {
-        let mut buf = Vec::new();
+        let payload = self.footer_payload()?;
+        let payload_size = payload.len();
+
+        let capacity = MIN_FOOTER_SIZE as usize + payload_size;
+        let mut buf = Vec::with_capacity(capacity);
+
         self.write_magic(&mut buf); // HeadMagic
-        let payload_size = self.write_footer_payload(&mut buf)?; // Payload
-        self.write_footer_payload_size(payload_size, &mut buf); // PayloadSize
+        buf.extend_from_slice(&payload); // Payload
+        self.write_footer_payload_size(payload_size as _, &mut buf); // PayloadSize
         self.write_flags(&mut buf); // Flags
         self.write_magic(&mut buf); // FootMagic
         Ok(buf)
@@ -54,18 +59,6 @@ impl FooterWriter {
 
     fn write_magic(&self, buf: &mut Vec<u8>) {
         buf.extend_from_slice(&MAGIC);
-    }
-
-    fn write_footer_payload(&mut self, buf: &mut Vec<u8>) -> Result<i32> {
-        let file_metadata = FileMetadataBuilder::default()
-            .blobs(mem::take(&mut self.blob_metadata))
-            .properties(mem::take(&mut self.file_properties))
-            .build()
-            .expect("Required fields are not set");
-
-        let json_data = serde_json::to_vec(&file_metadata).context(SerializeJsonSnafu)?;
-        buf.extend_from_slice(&json_data);
-        Ok(json_data.len() as i32)
     }
 
     fn write_footer_payload_size(&self, payload_size: i32, buf: &mut Vec<u8>) {
@@ -77,5 +70,15 @@ impl FooterWriter {
     /// TODO(zhongzc): support compression
     fn write_flags(&self, buf: &mut Vec<u8>) {
         buf.extend_from_slice(&[0; 4]);
+    }
+
+    fn footer_payload(&mut self) -> Result<Vec<u8>> {
+        let file_metadata = FileMetadataBuilder::default()
+            .blobs(mem::take(&mut self.blob_metadata))
+            .properties(mem::take(&mut self.file_properties))
+            .build()
+            .expect("Required fields are not set");
+
+        serde_json::to_vec(&file_metadata).context(SerializeJsonSnafu)
     }
 }
